# api/openapi.yaml
openapi: 3.1.0
info:
  title: buffgo local control API
  version: 0.1.0
servers:
  - url: /
paths:
  /api/security/context:
    get:
      operationId: getSecurityContext
      summary: Create a short-lived browser control session
      responses:
        "200":
          description: Security context created without changing business data
          headers:
            Set-Cookie:
              description: Host-only HttpOnly SameSite=Strict control-session cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityContext"
        "421":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "405":
          $ref: "#/components/responses/Error"
        "415":
          $ref: "#/components/responses/Error"
...
  /api/accounts:
    get:
      operationId: listAccounts
      security: []
      responses:
        "200":
          description: Credential-safe account summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountCreate"
      responses:
        "201":
          description: Account created; session is never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
  /api/accounts/{id}/alias:
    parameters:
      - $ref: "#/components/parameters/AccountID"
    post:
      operationId: renameAccount
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountAliasReplace"
      responses:
        "200":
          description: Alias updated; session is never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/accounts/{id}/session:
    parameters:
      - $ref: "#/components/parameters/AccountID"
    post:
      operationId: replaceAccountSession
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountSessionReplace"
      responses:
        "200":
          description: Session replaced; raw session is never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/accounts/{id}/delete:
    parameters:
      - $ref: "#/components/parameters/AccountID"
    post:
      operationId: deleteAccount
      security:
        - controlSession: []
          csrfToken: []
      responses:
        "200":
          description: Account deleted
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/nodes:
    get:
      operationId: listNodes
      responses:
        "200":
          description: Credential-safe access nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Node"
    post:
      operationId: createNode
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeCreate"
      responses:
        "201":
          description: Node created in validating state; proxy credential is never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
  /api/combinations:
    get:
      operationId: listCombinations
      responses:
        "200":
          description: Account-node combinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Combination"
    post:
      operationId: createCombination
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CombinationCreate"
      responses:
        "201":
          description: Combination created; platform is derived from the account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Combination"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
  /api/watermarks:
    get:
      operationId: listWatermarks
      responses:
        "200":
          description: Short-pool minimums; usable is current short nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Watermark"
    post:
      operationId: setWatermark
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatermarkSet"
      responses:
        "200":
          description: Watermark updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watermark"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/providers:
    get:
      operationId: listProviders
      responses:
        "200":
          description: Credential-safe vendor summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
    post:
      operationId: createProvider
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderCreate"
      responses:
        "201":
          description: Provider created; credential is never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
  /api/targets:
    get:
      operationId: listTargets
      responses:
        "200":
          description: Summary collection targets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Target"
    post:
      operationId: createTarget
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetCreate"
      responses:
        "201":
          description: Target created or existing identical target returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Target"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
  /api/runs:
    get:
      operationId: listRuns
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
      responses:
        "200":
          description: Recent collection runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
...
  /api/quote-facets:
    get:
      operationId: listQuoteFacets
      description: Games and item types present in collected data, for filter options.
      parameters:
        - name: appid
          in: query
          description: Scopes item types to one game. Games are always listed in full.
          schema: {type: integer, format: int64, minimum: 1}
      responses:
        "200":
          description: Available filter options
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [appids, item_types]
                properties:
                  appids:
                    type: array
                    items: {type: integer, format: int64, minimum: 1}
                  item_types:
                    type: array
                    items: {type: string}
        "400": {$ref: "#/components/responses/Error"}
  /api/quotes:
    get:
      operationId: listQuotes
      parameters:
        - name: appid
          in: query
          schema: {type: integer, format: int64, minimum: 1}
        - name: platform
          in: query
          schema: {type: string}
        - name: side
          in: query
          schema: {type: string, enum: [bid, ask]}
        - name: keyword
          in: query
          description: Case-insensitive substring match on the product name.
          schema: {type: string, maxLength: 128}
        - name: item_type
          in: query
          schema: {type: string, maxLength: 128}
        - name: min_cents
          in: query
          description: Only products with a last present price at or above this value.
          schema: {type: integer, format: int64, minimum: 0}
        - name: max_cents
          in: query
          schema: {type: integer, format: int64, minimum: 0}
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_desc, price_asc, listings_desc, name]
            default: price_desc
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, maximum: 1000000, default: 0}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 200, default: 60}
      responses:
        "200":
          description: One page of latest attempts with the total match count
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [total, quotes]
                properties:
                  total: {type: integer, format: int64, minimum: 0}
                  quotes:
                    type: array
                    items:
                      $ref: "#/components/schemas/Quote"
components:
  parameters:
    AccountID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  securitySchemes:
    controlSession:
      type: apiKey
      in: cookie
      name: buffgo_control_session
    csrfToken:
      type: apiKey
      in: header
      name: X-Buffgo-CSRF
  schemas:
    SecurityContext:
      type: object
      additionalProperties: false
      required:
        - csrf_token
        - expires_at
      properties:
        csrf_token:
          type: string
          minLength: 43
          maxLength: 43
        expires_at:
          type: string
          format: date-time
...
    Account:
      type: object
      additionalProperties: false
      required: [id, platform, alias, session_state, session_revision]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        platform: {type: string}
        alias: {type: string}
        session_state: {type: string, enum: [unverified, valid, invalid]}
        session_revision: {type: integer, format: int64, minimum: 1}
        last_checked_at: {type: string, format: date-time}
    AccountCreate:
      type: object
      additionalProperties: false
      required: [platform, alias, session]
      properties:
        platform: {type: string}
        alias: {type: string}
        session: {type: string, writeOnly: true, minLength: 1, maxLength: 262144}
    AccountAliasReplace:
      type: object
      additionalProperties: false
      required: [alias]
      properties:
        alias: {type: string, minLength: 1, maxLength: 128}
    AccountSessionReplace:
      type: object
      additionalProperties: false
      required: [expected_revision, session]
      properties:
        expected_revision: {type: integer, format: int64, minimum: 1}
        session: {type: string, writeOnly: true, minLength: 1, maxLength: 262144}
    Node:
      type: object
      additionalProperties: false
      required: [id, name, kind, region, egress_mode, state, egress_revision, assignment_revision, sides, has_proxy_credential]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        name: {type: string}
        kind: {type: string, enum: [direct, proxy]}
        region: {type: string, enum: [domestic, foreign, hongkong]}
        egress_mode: {type: string, enum: [static, sticky]}
        state: {type: string, enum: [validating, available, unavailable]}
        egress_revision: {type: integer, format: int64, minimum: 1}
        assignment_revision: {type: integer, format: int64, minimum: 1}
        appid: {type: integer, format: int64, minimum: 1}
        sides:
          type: array
          items:
            $ref: "#/components/schemas/NodeSide"
        has_proxy_credential: {type: boolean}
        sticky_session_valid_until: {type: string, format: date-time}
        exit:
          $ref: "#/components/schemas/NodeExitView"
    NodeSide:
      type: object
      additionalProperties: false
      required: [platform, side]
      properties:
        platform: {type: string}
        side: {type: string, enum: [bid, ask]}
    NodeExitView:
      type: object
      additionalProperties: false
      required: [address, verified_at, valid_until]
      properties:
        address: {type: string}
        verified_at: {type: string, format: date-time}
        valid_until: {type: string, format: date-time}
    NodeCreate:
      type: object
      additionalProperties: false
      required: [name, kind, region, egress_mode]
      properties:
        name: {type: string}
        kind: {type: string, enum: [direct, proxy]}
        region: {type: string, enum: [domestic, foreign, hongkong]}
        egress_mode: {type: string, enum: [static, sticky]}
        sticky_session_valid_until: {type: string, format: date-time}
        proxy_credential: {type: string, writeOnly: true}
...
    Combination:
      type: object
      additionalProperties: false
      required: [id, platform, account_id, node_id]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        platform: {type: string}
        account_id: {type: integer, format: int64, minimum: 1}
        node_id: {type: integer, format: int64, minimum: 1}
    Watermark:
      type: object
      additionalProperties: false
      required: [region, min_usable, usable, revision]
      properties:
        region: {type: string, enum: [domestic, foreign, hongkong]}
        min_usable: {type: integer, minimum: 0}
        usable: {type: integer, minimum: 0}
        revision: {type: integer, format: int64, minimum: 1}
    WatermarkSet:
      type: object
      additionalProperties: false
      required: [region, expected_revision, min_usable]
      properties:
        region: {type: string, enum: [domestic, foreign, hongkong]}
        expected_revision: {type: integer, format: int64, minimum: 1}
        min_usable: {type: integer, minimum: 0}
    Provider:
      type: object
      additionalProperties: false
      required: [id, name, enabled, priority, regions, has_credential, revision]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        name: {type: string}
        enabled: {type: boolean}
        priority: {type: integer, minimum: 1, maximum: 10000}
        regions:
          type: array
          items: {type: string, enum: [domestic, foreign, hongkong]}
        has_credential: {type: boolean}
        revision: {type: integer, format: int64, minimum: 1}
    ProviderCreate:
      type: object
      additionalProperties: false
      required: [name, priority, regions, credential]
      properties:
        name: {type: string}
        enabled: {type: boolean}
        priority: {type: integer, minimum: 1, maximum: 10000}
        regions:
          type: array
          items: {type: string, enum: [domestic, foreign, hongkong]}
        credential: {type: string, writeOnly: true, minLength: 1, maxLength: 65536}
...
    CombinationCreate:
      type: object
      additionalProperties: false
      required: [account_id, node_id]
      properties:
        account_id: {type: integer, format: int64, minimum: 1}
        node_id: {type: integer, format: int64, minimum: 1}
...
    TargetCreate:
      type: object
      additionalProperties: false
      required: [platform, appid, side]
      properties:
        platform: {type: string}
        appid: {type: integer, format: int64, minimum: 1}
        side: {type: string, enum: [bid, ask]}
        desired: {type: string, enum: [enabled, disabled]}
...
  responses:
    Error:
      description: Request rejected without reflecting request data or credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"